refactor: extract lyric input area into standalone component
Extracts the lyric paste textbox, setting toggles (Word-by-word, Char-based, Duet), offset input, and upload/load lyric buttons from `App.vue` into a dedicated, modular Vue component `LyricInputArea.vue`. Resolves Phase 3 of the architectural roadmap.
